Create official version when draft has been reviewed.
